Strip arguments into multiple files so it is not that big anymore
